Refactored current code base to provide a more responsive layout.
