+++ react_front_end/src/pages/components/GameStatistics.scss
@@ -1,0 +1,36 @@
+.statistics {
+  &.left,
+  &.right {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    text-align: center;
+    font-size: 80%;
+    height: 50%;
+  }
+
+  &.column {
+    display: flex;
+    flex-direction: column;
+  }
+
+  &.row {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+
+    > div {
+      padding: 2em;
+    }
+
+    @media screen and (max-width: 600px) {
+      flex-wrap: wrap;
+      width: 100%;
+      height: 50%;
+      > div {
+        height: 2em;
+        width: 100%;
+      }
+    }
+  }
+}
