+++ react_front_end/src/pages/components/LogoContainer.scss
@@ -1,0 +1,18 @@
+.root {
+  display: flex;
+  background-color: black;
+  width: 100vw;
+  height: 100%;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+
+  @media screen and (max-width: 600px) {
+    .column {
+      flex-direction: column;
+    }
+    .row {
+      flex-direction: row;
+    }
+  }
+}
