+++ react_front_end/src/pages/components/TriangleLogo.scss
@@ -1,12 +1,38 @@
-.statistics {
+.triangle-logo {
+  min-height: 50vh;
 
-  &.left,
-  &.right {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    text-align: center;
+  animation-name: scale-triangle;
+  animation-duration: 2s;
+  animation-fill-mode: forwards;
+  animation-iteration-count: 1;
 
-    font-size: 80%;
+  @keyframes scale-triangle {
+    0% {
+      transform: scale(0, 0);
+    }
+    100% {
+      transform: scale(1, 1);
+    }
+  }
+
+  @media screen and (min-width: 0px) {
+    // background-color: orange;
+    width: 50vw;
+  }
+  @media screen and (min-width: 600px) {
+    // background-color: red;
+    max-width: 45vw;
+  }
+  @media screen and (min-width: 960px) {
+    // background-color: green;
+    max-width: 30vw;
+  }
+  @media screen and (min-width: 1280px) {
+    // background-color: blue;
+    max-width: 25vw;
+  }
+  @media screen and (min-width: 1920px) {
+    // background-color: purple;
+    max-width: 20vw;
   }
 }
